+++ packages/nameai-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @namehash/nameai
 
+## 1.1.2
+
+### Patch Changes
+
+- eaa085e: Refine readme
+
+## 1.1.1
+
+### Patch Changes
+
+- 08c5ae0: Refine NameAI Readme Files
+
 ## 1.1.0
 
 ### Minor Changes